requirements.txt
absl-py==2.3.1
actionlib==1.14.3
angles==1.9.14
ansitable==0.11.4
asttokens==3.0.0
astunparse==1.6.3
backcall==0.2.0
bondpy==1.8.7
cachetools==5.5.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
certifi==2025.7.14
cfgv==3.4.0
charset-normalizer==3.4.2
colored==1.4.4
comm==0.2.3
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
distlib==0.4.0
distro==1.9.0
docopt==0.6.2
docutils==0.20.1
dynamic-reconfigure==1.7.6
empy==4.2
etils==1.3.0
evdev==1.9.2
executing==2.2.0
feetech-servo-sdk==1.0.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.4.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
glfw==2.9.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
hid==1.0.4
hidapi==0.14.0.post4
identify==2.6.1
idna==3.10
image-geometry==1.16.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
interactive-markers==1.12.2
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.13.1
kiwisolver==1.4.7
laser-geometry==1.6.8
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
message-filters==1.17.4
nest-asyncio==1.6.0
nodeenv==1.9.1
num2words==0.5.14
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pgraph-python==0.6.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pre-commit==3.5.0
progress==1.6.1
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat==0.6.6
pyglm==2.8.2
Pygments==2.19.2
pynput==1.7.7
PyOpenGL==3.1.9
pyparsing==3.1.4
pyrealsense2==2.55.1.6486
pyserial==3.5
python-dateutil==2.9.0.post0
python-qt-binding==0.4.6
python-xlib==0.33
pyttsx3==2.99
PyYAML==6.0.2
pyzmq==27.0.1
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
requests==2.32.4
requests-oauthlib==2.0.0
resource-retriever==1.12.10
rlds==0.1.8
roboticstoolbox-python==1.1.1
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rsa==4.9.1
rtb-data==1.0.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.8
rviz==1.14.26
scipy==1.10.1
sensor-msgs==1.13.2
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
spatialgeometry==1.1.0
spatialmath-python==1.1.14
stack-data==0.6.3
swift-sim==1.1.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
topic-tools==1.17.4
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.5.0
urllib3==2.2.3
virtualenv==20.33.1
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.2
xacro==1.14.20
xarm-python-sdk==1.15.3
zipp==3.20.2
